updated analysis script and bat file
The analysis script has changed the column positions of the mks vdata due to a change in data log column size. The bat file has also been updated to use powershell explicitly, this has removed the terminal prompt that had been appearing before due to the size of the raw data file.